+++ src/modules/enviroment/interface/environment-team.controller.ts
@@ -9,12 +9,13 @@ import {
   Query,
   UseGuards,
 } from '@nestjs/common';
+import { ApiTags } from '@nestjs/swagger';
 
 import { IPromiseResponse } from '@/common/response_service/interface/response.interface';
 import { Auth } from '@/modules/auth/application/decorator/auth.decorator';
 import { AuthType } from '@/modules/auth/domain/auth_type.enum';
-import { AdminRoleGuard } from '@/modules/authorization/infraestructure/policy/guard/admin-role.guard';
-import { AuthTeamGuard } from '@/modules/authorization/infraestructure/policy/guard/auth-team.guard';
+import { AdminRoleGuard } from '@/modules/authorization/guard/admin-role.guard';
+import { AuthTeamGuard } from '@/modules/authorization/guard/auth-team.guard';
 
 import { CreateEnviromentDto } from '../application/dto/create-enviroment.dto';
 import { EnviromentResponseDto } from '../application/dto/enviroment-response.dto';
@@ -22,6 +23,7 @@ import { UpdateEnviromentDto } from '../application/dto/update-enviroment.dto';
 import { EnviromentService } from '../application/service/enviroment.service';
 
 @Auth(AuthType.Bearer)
+@ApiTags('Environment Team')
 @UseGuards(AuthTeamGuard)
 @Controller('/team/:teamId/environment')
 export class EnviromentTeamController {
